+++ src/main/java/me/dio/api_rest_springboot/domain/model/Planeta.java
@@ -13,17 +13,17 @@ public class Planeta {
     
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
-    private Double id;
+    private Integer id;
 
     @Column (unique = true)
     private String nome;
     private Double tamanho;
     // private Galaxia galaxia;
 
-    public Double getId() {
+    public Integer getId() {
         return id;
     }
-    public void setId(Double id) {
+    public void setId(Integer id) {
         this.id = id;
     }
     public String getNome() {
