+++ src/main/java/me/dio/api_rest_springboot/domain/repository/EstrelaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import me.dio.api_rest_springboot.domain.model.Estrela;
 
 @Repository
-public interface EstrelaRepository extends JpaRepository<Estrela, Double> {
+public interface EstrelaRepository extends JpaRepository<Estrela, Integer> {
 
 
 }
